extras/requirements.txt
# Core requirements
openai>=1.0.0
python-dotenv>=1.0.0
azure-cognitiveservices-speech>=1.31.0
PyQt6>=6.4.0

# Transcription extras
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0
faster-whisper>=0.10.0 